# docs/swagger.yaml
basePath: /
definitions:
...
  GetSummaryByDrugIDRow:
    properties:
      adverse_effects:
        type: string
      contraindications:
        type: string
      drug_id:
        type: integer
      drug_interactions:
        type: string
      id:
        type: integer
      leaflet_expedient:
        type: string
      leaflet_published_at:
        type: string
      mechanism_of_action:
        type: string
      missed_dose:
        type: string
      posology:
        type: string
      reviewed_at:
        type: string
      reviewed_by:
        type: string
      side_effects:
        type: string
      source_url:
        type: string
      storage:
        type: string
      updated_at:
        type: string
      warnings:
        type: string
      what_is_it_for:
        type: string
      when_to_seek_help:
        type: string
    type: object
...
  GetSummaryByIDRow:
    properties:
      adverse_effects:
        type: string
      contraindications:
        type: string
      drug_id:
        type: integer
      drug_interactions:
        type: string
      id:
        type: integer
      leaflet_expedient:
        type: string
      leaflet_published_at:
        type: string
      mechanism_of_action:
        type: string
      missed_dose:
        type: string
      posology:
        type: string
      reviewed_at:
        type: string
      reviewed_by:
        type: string
      side_effects:
        type: string
      source_url:
        type: string
      storage:
        type: string
      updated_at:
        type: string
      warnings:
        type: string
      what_is_it_for:
        type: string
      when_to_seek_help:
        type: string
    type: object
...
  ListSummariesRow:
    properties:
      active_ingredient:
        type: string
      brand_name:
        type: string
      drug_id:
        type: integer
      id:
        type: integer
      leaflet_expedient:
        type: string
      leaflet_published_at:
        type: string
      reviewed_at:
        type: string
      reviewed_by:
        type: string
      source_url:
        type: string
      updated_at:
        type: string
      what_is_it_for:
        type: string
    type: object
...
  createSummaryRequest:
    properties:
      adverse_effects:
        type: string
      contraindications:
        type: string
      drug_id:
        type: integer
      drug_interactions:
        type: string
      leaflet_expedient:
        description: Version of the official leaflet that was summarized
        example: 0123456/24-5
        type: string
      leaflet_published_at:
        example: "2024-05-31"
        format: date
        type: string
      mechanism_of_action:
        type: string
      missed_dose:
        type: string
      posology:
        type: string
      side_effects:
        type: string
      source_url:
        type: string
      storage:
        type: string
      warnings:
        type: string
      what_is_it_for:
        type: string
      when_to_seek_help:
        type: string
    type: object
...
  idResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  listResponse:
    properties:
      data: {}
      limit:
        type: integer
      offset:
        type: integer
    type: object
...
  updateSummaryRequest:
    properties:
      adverse_effects:
        type: string
      contraindications:
        type: string
      drug_interactions:
        type: string
      leaflet_expedient:
        description: Version of the official leaflet that was summarized
        example: 0123456/24-5
        type: string
      leaflet_published_at:
        example: "2024-05-31"
        format: date
        type: string
      mechanism_of_action:
        type: string
      missed_dose:
        type: string
      posology:
        type: string
      side_effects:
        type: string
      source_url:
        type: string
      storage:
        type: string
      warnings:
        type: string
      what_is_it_for:
        type: string
      when_to_seek_help:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    API that reads the barcode on a medicine box and returns the simplified leaflet.
    Public route: GET /drugs/ean/{ean}. All others require login (POST /auth/login sets a session cookie).
  title: PharmaCode API
  version: "1.0"
paths:
...
  /summaries:
    get:
      parameters:
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: How many items to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/listResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ListSummariesRow'
                  type: array
              type: object
        "400":
          description: invalid limit / invalid offset
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: List simplified leaflets (paginated)
      tags:
      - summaries
    post:
      consumes:
      - application/json
      parameters:
      - description: Leaflet summary
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/createSummaryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/idResponse'
        "400":
          description: invalid json or missing required field
          schema:
            type: string
        "404":
          description: drug not found for this drug_id
          schema:
            type: string
        "409":
          description: this drug already has a summary
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create the simplified leaflet of a drug
      tags:
      - summaries
  /summaries/{id}:
    delete:
      parameters:
      - description: Summary ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id
          schema:
            type: string
        "404":
          description: summary not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete a simplified leaflet
      tags:
      - summaries
    get:
      parameters:
      - description: Summary ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetSummaryByIDRow'
        "400":
          description: invalid id
          schema:
            type: string
        "404":
          description: summary not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get a simplified leaflet by ID
      tags:
      - summaries
    put:
      consumes:
      - application/json
      parameters:
      - description: Summary ID
        in: path
        name: id
        required: true
        type: integer
      - description: Leaflet summary
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/updateSummaryRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id or json / missing required field
          schema:
            type: string
        "404":
          description: summary not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Update a simplified leaflet
      tags:
      - summaries
  /summaries/{id}/review:
    patch:
      description: The reviewer is the logged-in user (reviewer or admin role). No
        request body.
      parameters:
      - description: Summary ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id
          schema:
            type: string
        "401":
          description: authentication required
          schema:
            type: string
        "403":
          description: insufficient permissions
          schema:
            type: string
        "404":
          description: summary not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Mark a simplified leaflet as reviewed
      tags:
      - summaries
  /summaries/drug/{drugID}:
    get:
      parameters:
      - description: Drug ID
        in: path
        name: drugID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetSummaryByDrugIDRow'
        "400":
          description: invalid drug id
          schema:
            type: string
        "404":
          description: summary not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get the simplified leaflet of a drug
      tags:
      - summaries
...
swagger: "2.0"